requirements.txt
python-dotenv==1.0.0
telepot==12.7
